Pluggable Containerizer
This patch adds the so-called pluggable containerizer. This
containerizer delegates all containerizer calls directly to
an external containerizer program (which can be specified on
start-up). Few calls have internal fall-back implementations
such as wait(), destroy() and usage().
The protocol for the interactions with the external program
is as follows:
./containerizer <command> <container-id> (<optional parameters>)
When protocol buffers need to be provided, the Mesos side of
the pluggable containerizer will serialize the protos on stdin
and vice-versa for reading protos on stdout.

NOTE: Still work in progress and not yet reviewable.

Review: https://reviews.apache.org/r/17567