Added support for passing s3[n]:// URIs to hadoop
The fetcher previously would only call out to the hadoop client when it encountered a hdfs:// or htfp:// URI meaning Mesos has no support for Amazon S3. Hadoop already has built-in implementations of S3 (both as a block store and as a K/V store) so this commit simply changes the fetcher to pass those schemes through to hadoop too.